# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FAM Backend API",
    "version": "1.0.0",
    "description": "Family event management platform — three-role auth (user/admin/organizer), family tree, QR join, notifications."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "EventObject": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "64b1f2c3e4d5f6a7b8c9d0e1" },
          "name": { "type": "string", "example": "Smith Wedding" },
          "code": { "type": "string", "example": "smithweddi4821" },
          "qrCode": { "type": "string", "example": "https://res.cloudinary.com/demo/image/upload/v1/qr/event_64b1f2c3.png" },
          "eventDate": { "type": "string", "format": "date-time" },
          "approvalMode": { "type": "string", "enum": ["auto", "manual"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/events": {
      "post": {
        "summary": "Create a new event — automatically creates an Organizer account (temp password emailed)",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventName", "eventDate", "organizerEmail", "organizerName"],
                "properties": {
                  "eventName": { "type": "string", "example": "Smith Wedding" },
                  "organizerEmail": { "type": "string", "example": "org@example.com" },
                  "organizerName": { "type": "string", "example": "Sarah Jones" },
                  "description": { "type": "string", "example": "Annual family wedding celebration" },
                  "eventType": {
                    "type": "string",
                    "enum": ["wedding", "reunion", "birthday", "funeral", "festival", "other"],
                    "default": "other"
                  },
                  "eventDate": { "type": "string", "format": "date", "example": "2026-06-15" },
                  "eventEndDate": { "type": "string", "format": "date", "example": "2026-06-16" },
                  "approvalMode": { "type": "string", "enum": ["auto", "manual"], "default": "manual" },
                  "settings": {
                    "type": "object",
                    "properties": {
                      "approvalRequired": { "type": "boolean" },
                      "maxMembersPerUser": { "type": "integer", "example": 4 }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event and organizer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Event created successfully" },
                    "event": { "$ref": "#/components/schemas/EventObject" },
                    "organizer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "tempPassword": { "type": "string", "example": "abc12345" },
                        "accessCode": { "type": "string", "example": "XY12AB" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing required fields" }
        }
      },
      "get": {
        "summary": "Paginated list of all events",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 } },
          { "in": "query", "name": "isActive", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "Paginated array of event objects" }
        }
      }
    },
...
    "/api/admin/events/{id}": {
      "delete": {
        "summary": "Soft-delete an event (sets isActive=false). Also deactivates all organizers for the event.",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event deleted" },
          "404": { "description": "Event not found" }
        }
      }
    },
...
    "/api/admin/reports/event/{eventId}": {
      "get": {
        "summary": "Reports for a specific event",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["pending", "reviewed", "dismissed", "actioned"] } },
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Reports for the event" }
        }
      }
    },
...
    "/api/events/validate-code": {
      "post": {
        "summary": "Check if an event code is valid before joining (public, rate limited)",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventCode"],
                "properties": {
                  "eventCode": { "type": "string", "example": "smithweddi4821" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event found — returns event summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "event": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "description": { "type": "string" },
                        "eventDate": { "type": "string", "format": "date-time" },
                        "requiresApproval": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Invalid or inactive event code" }
        }
      }
    },
    "/api/events/join": {
      "post": {
        "summary": "Join an event by code. Routes to participants (auto) or pendingApprovals (manual).",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventCode"],
                "properties": {
                  "eventCode": { "type": "string", "example": "smithweddi4821" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "joinStatus is approved (auto mode) or pending_approval (manual mode)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "event": { "$ref": "#/components/schemas/EventObject" },
                    "requiresApproval": { "type": "boolean" },
                    "joinStatus": { "type": "string", "enum": ["approved", "pending_approval"] }
                  }
                }
              }
            }
          },
          "400": { "description": "Already joined or invalid code" }
        }
      }
    },
    "/api/events": {
      "get": {
        "summary": "Paginated event list for the authenticated user",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Array of events" }
        }
      }
    },
    "/api/events/{id}": {
      "get": {
        "summary": "Single event with populated creator, organizers and participants",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event object" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "summary": "Update event fields (admin only)",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventName": { "type": "string" },
                  "description": { "type": "string" },
                  "eventDate": { "type": "string", "format": "date" },
                  "eventEndDate": { "type": "string", "format": "date" },
                  "eventType": { "type": "string", "enum": ["wedding", "reunion", "birthday", "funeral", "festival", "other"] },
                  "approvalMode": { "type": "string", "enum": ["auto", "manual"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Event updated" }
        }
      }
    },
    "/api/events/{id}/qr": {
      "get": {
        "summary": "Get QR code image URL for an event",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "QR code Cloudinary URL and event code",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "qrCode": { "type": "string", "example": "https://res.cloudinary.com/demo/image/upload/v1/qr/event_64b1.png" },
                    "eventCode": { "type": "string", "example": "smithweddi4821" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{eventId}/participants": {
      "get": {
        "summary": "Paginated, searchable participant list. Users must be event participants; org/admin have unrestricted access.",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "search", "schema": { "type": "string" }, "description": "Search by username or email" },
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Array of participant users" }
        }
      }
    },
...
    "/api/reports/event/{eventId}": {
      "get": {
        "summary": "Reports for a specific event (organizer or admin)",
        "tags": ["Reports"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["pending", "reviewed", "dismissed", "actioned"] } },
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Reports for the event" }
        }
      }
    },
...
  },
...
}